+++ pkg/event_metadata_renamer/renamer.go
@@ -1,4 +1,4 @@
-package event_metadata_rename
+package event_metadata_renamer
 
 import (
 	"fmt"
@@ -21,33 +21,33 @@ var (
 	}, []string{"Source", "Destination"})
 )
 
-func NewFromViper(viperConfig *viper.Viper, logger logrus.FieldLogger) (*eventMetadataRenameManager, error) {
-	var renameConfig []renameConfig
-	marshalledConfig, err := yaml.Marshal(viperConfig.Get("eventMetadataRenameConfigs"))
+func NewFromViper(viperConfig *viper.Viper, logger logrus.FieldLogger) (*eventMetadataRenamerManager, error) {
+	var config []renamerConfig
+	marshalledConfig, err := yaml.Marshal(viperConfig.Get("eventMetadataRenamerConfigs"))
 	if err != nil {
 		return nil, fmt.Errorf("failed to load configuration: %w", err)
 	}
-	if err := yaml.UnmarshalStrict(marshalledConfig, &renameConfig); err != nil {
+	if err := yaml.UnmarshalStrict(marshalledConfig, &config); err != nil {
 		return nil, fmt.Errorf("failed to load configuration: %w", err)
 	}
-	return NewFromConfig(renameConfig, logger)
+	return NewFromConfig(config, logger)
 }
 
 // New returns requestNormalizer which allows to add Key to RequestEvent
-func NewFromConfig(renameConfig []renameConfig, logger logrus.FieldLogger) (*eventMetadataRenameManager, error) {
-	relabelManager := eventMetadataRenameManager{
-		renameConfig: renameConfig,
+func NewFromConfig(config []renamerConfig, logger logrus.FieldLogger) (*eventMetadataRenamerManager, error) {
+	relabelManager := eventMetadataRenamerManager{
+		renamerConfig: config,
 		outputChannel: make(chan *event.Raw),
 		logger:        logger,
 	}
 	return &relabelManager, nil
 }
-type renameConfig struct {
+type renamerConfig struct {
 	Source, Destination string
 }
 
-type eventMetadataRenameManager struct {
-	renameConfig []renameConfig
+type eventMetadataRenamerManager struct {
+	renamerConfig []renamerConfig
 	observer      pipeline.EventProcessingDurationObserver
 	inputChannel  chan *event.Raw
 	outputChannel chan *event.Raw
@@ -55,50 +55,49 @@ type eventMetadataRenameManager struct {
 	logger        logrus.FieldLogger
 }
 
-func (r *eventMetadataRenameManager) String() string {
-	return "eventMetadataRename"
+func (r *eventMetadataRenamerManager) String() string {
+	return "eventMetadataRenamer"
 }
 
-func (r *eventMetadataRenameManager) Done() bool {
+func (r *eventMetadataRenamerManager) Done() bool {
 	return r.done
 }
 
-func (r *eventMetadataRenameManager) RegisterMetrics(_ prometheus.Registerer, wrappedRegistry prometheus.Registerer) error {
+func (r *eventMetadataRenamerManager) RegisterMetrics(_ prometheus.Registerer, wrappedRegistry prometheus.Registerer) error {
 	return wrappedRegistry.Register(renamingCollisionsTotal)
 }
 
-func (r *eventMetadataRenameManager) SetInputChannel(channel chan *event.Raw) {
+func (r *eventMetadataRenamerManager) SetInputChannel(channel chan *event.Raw) {
 	r.inputChannel = channel
 }
 
-func (r *eventMetadataRenameManager) OutputChannel() chan *event.Raw {
+func (r *eventMetadataRenamerManager) OutputChannel() chan *event.Raw {
 	return r.outputChannel
 }
 
-func (r *eventMetadataRenameManager) Stop() {
+func (r *eventMetadataRenamerManager) Stop() {
 	return
 }
 
-func (r *eventMetadataRenameManager) RegisterEventProcessingDurationObserver(observer pipeline.EventProcessingDurationObserver) {
+func (r *eventMetadataRenamerManager) RegisterEventProcessingDurationObserver(observer pipeline.EventProcessingDurationObserver) {
 	r.observer = observer
 }
 
-func (r *eventMetadataRenameManager) observeDuration(start time.Time) {
+func (r *eventMetadataRenamerManager) observeDuration(start time.Time) {
 	if r.observer != nil {
 		r.observer.Observe(time.Since(start).Seconds())
 	}
 }
 
 // renameEventMetadata applies the relabel configs on the event metadata.
-func (r *eventMetadataRenameManager) renameEventMetadata(e *event.Raw) *event.Raw {
-	// TODO
-	for _, renameConfig := range r.renameConfig {
+func (r *eventMetadataRenamerManager) renameEventMetadata(e *event.Raw) *event.Raw {
+	for _, renameConfig := range r.renamerConfig {
+		if _, ok := e.Metadata[renameConfig.Source]; !ok {
+			continue
+		}
 		if _, ok := e.Metadata[renameConfig.Destination]; ok {
 			r.logger.Warnf("refusing to override metadata's %s:%s with %s:%s", renameConfig.Destination, e.Metadata[renameConfig.Destination], renameConfig.Source, e.Metadata[renameConfig.Source])
 			renamingCollisionsTotal.WithLabelValues(renameConfig.Source, renameConfig.Destination).Inc()
-			continue
-		}
-		if _, ok := e.Metadata[renameConfig.Source]; !ok {
 			continue
 		}
 		e.Metadata[renameConfig.Destination] = e.Metadata[renameConfig.Source]
@@ -108,7 +107,7 @@ func (r *eventMetadataRenameManager) renameEventMetadata(e *event.Raw) *event.Ra
 }
 
 // Run event replacer receiving events and filling their Key if not already filled.
-func (r *eventMetadataRenameManager) Run() {
+func (r *eventMetadataRenamerManager) Run() {
 	go func() {
 		defer func() {
 			close(r.outputChannel)